Fixed PDF report
