+++ src/pyNAVIS/__init__.py
@@ -1,6 +1,6 @@
 from .dataset_gen import DatasetGenerators
 from .functions import Functions
-from .functions import ReportFunctions
+from .report_functions import ReportFunctions
 from .generators import Generators
 from .loaders import Loaders
 from .objects import SpikesFile
